restructured the initialization of shad
